Conflict resolutions:
- package.json: kept both sides' scripts (check:route-contract +
  main's check:drizzle-config / check:schema-builders).
- usePdmEquipmentDetailData.ts: both sides independently made the same
  raw-fetch → apiRequest fix; kept this branch's explanatory comment.
- pdm-equipment-detail.tsx: took main's stricter healthScore null guard
  (functionally equivalent).
- use-upload.ts (modify/delete): deletion kept — main's change was the
  mechanical WS4 envelope sweep and the module still has zero importers
  on the merged tree (knip-confirmed dead, per the route-contract triage).

Route-contract reconciliation on the merged tree: main's work resolved
/api/pdm/risk and /api/optimization/trend-insights (pruned); two new
entries baselined with documented false-positive dispositions
(/api/observability is errorHandler's self-log classifier pattern;
/api/optimization is a deliberate queryKey base for prefix invalidation
— the real fetch /api/optimization/dashboard exists). Baseline: 130.

+++ server/purchasing/po-routes.ts
@@ -80,13 +80,9 @@ router.get("/", requireOrgId, generalLimit, async (req, res) => {
         updatedAt: purchaseOrders.updatedAt,
         supplierName: suppliers.name,
         supplierCode: suppliers.code,
-        totalQty: sql<number>`COALESCE(SUM(${purchaseOrderItems.quantity}), 0)`.as("totalQty"),
-        receivedQty: sql<number>`COALESCE(SUM(${purchaseOrderItems.receivedQuantity}), 0)`.as(
-          "receivedQty"
-        ),
-        rejectedQty: sql<number>`COALESCE(SUM(${purchaseOrderItems.rejectedQuantity}), 0)`.as(
-          "rejectedQty"
-        ),
+        totalQty: sql<number>`COALESCE(SUM(${purchaseOrderItems.quantity}), 0)::float8`.as("totalQty"),
+        receivedQty: sql<number>`COALESCE(SUM(${purchaseOrderItems.receivedQuantity}), 0)::float8`.as("receivedQty"),
+        rejectedQty: sql<number>`COALESCE(SUM(${purchaseOrderItems.rejectedQuantity}), 0)::float8`.as("rejectedQty"),
       })
       .from(purchaseOrders)
       .leftJoin(suppliers, eq(purchaseOrders.supplierId, suppliers.id))
